fix: cover React use alias edge cases
